Matrix reconciliation complete
Matches groups larger than singles, to remove from other non-matching members of grid. This logic is not enough to solve extreme puzzles. Incorporporating more advanced techniques such as "candidate line" to help with those. Also changed the grid class to distinguish whether or not it is a square or line grid.