Higher level logic started: Candidate line test
(If a number can only appear in 2 or 3 places in the same row or column in a nine-square grid, you need to remove them from anywhere else in the corresponding nine-square row or column. And vice versa.

The tests for this are complete and fire off a console.log. Next step is to have it package up the instructions for other grids. Then we'll have the other grids catch those instructions.