Candidate line test complete - now solves some (but not all) extreme diff. puzzles
if a grid's possibility matrix knows its possibilities are arranged in a way that theis rule applies, it tells the puzzleboard to cycle through the grids until it finds the other one (we know it's the other one because it's type is a grid where this is a line, or vice versa) holding the same squares, and then it tells that grid to remove that value from all of the other squares in the grid.

However, this isn;'t the only test in question. will need to apply more tests/logic.